Move ising translators out of cplex algo folder
