+++ qiskit_aqua/algorithms/translators/ising/clique.py
