+++ qiskit_aqua/algorithms/translators/ising/graphpartition.py
