+++ qiskit_aqua/algorithms/translators/ising/maxcut.py
