+++ qiskit_aqua/algorithms/translators/ising/partition.py
