+++ qiskit_aqua/algorithms/translators/ising/stableset.py
