+++ qiskit_aqua/algorithms/translators/ising/tsp.py
